78 spring data plusz gyakorlasa

+++ src/main/java/hu/webuni/hr/dodi/repository/PositionRepository.java
@@ -6,7 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import hu.webuni.hr.dodi.model.Position;
 
-public interface PositionRepository extends JpaRepository<Position, Integer> {
+public interface PositionRepository extends JpaRepository<Position, Long> {
 
 	public List<Position> findByName(String name);
+
 }
